feat: add delete method with manual cascade deletion for user

+++ src/domain/repositories/user/userRepository.ts
@@ -3,6 +3,7 @@ import { UserRepositoryInterface } from './userRepository.Interface'
 
 import prisma from '../../../infrastructure/prisma/prisma'
 import { TuserUpdate } from '@/domain/types/user'
+import AppError from '../../../domain/error/appError'
 
 export class UserRepository implements UserRepositoryInterface {
   async findByEmail(email: string): Promise<User | null> {
@@ -45,8 +46,80 @@ export class UserRepository implements UserRepositoryInterface {
   }
 
   async delete(id: string): Promise<void> {
-    await prisma.user.delete({
+    const user = await prisma.user.findUnique({
       where: { id },
+      include: {
+        healthUnit: {
+          include: {
+            equipment: true,
+            reports: true,
+          },
+        },
+        company: {
+          include: {
+            healthUnits: {
+              include: {
+                equipment: true,
+                reports: true,
+              },
+            },
+            equipment: true,
+            reports: true,
+          },
+        },
+      },
+    })
+
+    if (!user) {
+      throw new AppError('User not found', 404)
+    }
+
+    await prisma.$transaction(async prisma => {
+      if (user.healthUnit) {
+        await prisma.equipment.deleteMany({
+          where: { healthUnitId: user.healthUnit.id },
+        })
+
+        await prisma.report.deleteMany({
+          where: { healthUnitId: user.healthUnit.id },
+        })
+
+        await prisma.healthUnit.delete({
+          where: { id: user.healthUnit.id },
+        })
+      }
+
+      if (user.company) {
+        await prisma.equipment.deleteMany({
+          where: { companyId: user.company.id },
+        })
+
+        await prisma.report.deleteMany({
+          where: { companyId: user.company.id },
+        })
+
+        for (const healthUnit of user.company.healthUnits) {
+          await prisma.equipment.deleteMany({
+            where: { healthUnitId: healthUnit.id },
+          })
+
+          await prisma.report.deleteMany({
+            where: { healthUnitId: healthUnit.id },
+          })
+
+          await prisma.healthUnit.delete({
+            where: { id: healthUnit.id },
+          })
+        }
+
+        await prisma.company.delete({
+          where: { id: user.company.id },
+        })
+      }
+
+      await prisma.user.delete({
+        where: { id },
+      })
     })
   }
 }
